replace the rest of Vector with std::vector, now only VectorProtected remains, which has one functional change: [] will not do 2 critical sections anymore (as it called elementat before)
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3578 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ propbrowser.cpp
@@ -287,7 +287,7 @@ void SmartBrowser::CreateFromDispatch(HWND hwndParent, VectorProtected<ISelect> 
    {
       PropertyPane * const pproppane = m_vproppane[i];
       ExpandoInfo * const pexinfo = new ExpandoInfo();
-      pexinfo->m_id = i;
+      pexinfo->m_id = (int)i;
       pexinfo->m_fExpanded = false;
       pexinfo->m_psb = this;
